requirements.txt
requests 
ConfigParser
google-cloud-storage==1.44.0
ndjson